+++ src/main/java/edu/tamu/ctv/entity/Importinfo.java
@@ -9,7 +9,7 @@ import java.util.Date;
 public class Importinfo implements java.io.Serializable
 {
 
-	private int id;
+	private Long id;
 	private Users users;
 	private Date transactiondate;
 	private String notes;
@@ -20,7 +20,7 @@ public class Importinfo implements java.io.Serializable
 	{
 	}
 
-	public Importinfo(int id, Users users, Date transactiondate, String notes, byte[] file, String filename)
+	public Importinfo(Long id, Users users, Date transactiondate, String notes, byte[] file, String filename)
 	{
 		this.id = id;
 		this.users = users;
@@ -30,12 +30,12 @@ public class Importinfo implements java.io.Serializable
 		this.filename = filename;
 	}
 
-	public int getId()
+	public Long getId()
 	{
 		return this.id;
 	}
 
-	public void setId(int id)
+	public void setId(Long id)
 	{
 		this.id = id;
 	}
