+++ src/main/java/edu/tamu/ctv/entity/Orders.java
@@ -7,7 +7,7 @@ package edu.tamu.ctv.entity;
 public class Orders implements java.io.Serializable
 {
 
-	private int id;
+	private Long id;
 	private Rowheaders rowheaders;
 	private int orderId;
 
@@ -15,19 +15,19 @@ public class Orders implements java.io.Serializable
 	{
 	}
 
-	public Orders(int id, Rowheaders rowheaders, int orderId)
+	public Orders(Long id, Rowheaders rowheaders, int orderId)
 	{
 		this.id = id;
 		this.rowheaders = rowheaders;
 		this.orderId = orderId;
 	}
 
-	public int getId()
+	public Long getId()
 	{
 		return this.id;
 	}
 
-	public void setId(int id)
+	public void setId(Long id)
 	{
 		this.id = id;
 	}
